requirements.txt
streamlit
yt-dlp
openai-whisper
transformers
torch
sentencepiece